Include nbval for automated testing of notebooks
Previously, there were no automated checks to ensure that, as our
codebases change, the output from out notebooks are what we expect.

This PR incorporates the pytest plugin `nbval` to run notebooks, check
their output, and validate that the expected output is generated.
A separate environment yml file is included for automated testing to
use.